Reworked tests organisation
