+++ src/test/java/visitor/TestEvaluator.java
@@ -1,7 +1,9 @@
-package calculator;
+package visitor;
 
 //Import Junit5 libraries for unit testing:
 
+import calculator.*;
+import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
@@ -27,7 +29,7 @@ class TestEvaluator {
 
     @Test
     void testEvaluatorMyNumber() {
-        assertEquals(new MyNumber(value1), calc.eval(new MyNumber(value1)));
+        Assertions.assertEquals(new MyNumber(value1), calc.eval(new MyNumber(value1)));
     }
 
     @ParameterizedTest
