+++ src/test/java/visitor/TestPrinter.java
@@ -1,5 +1,6 @@
-package calculator;
+package visitor;
 
+import calculator.*;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
